+++ Bot/Bot/DataSources/Alpaca/Models/Bar.cs
@@ -10,16 +10,16 @@ public class Bar
     public string T { get; set; }
 
     [JsonPropertyName("o")]
-    public decimal O { get; set; }
+    public double O { get; set; }
 
     [JsonPropertyName("h")]
-    public decimal H { get; set; }
+    public double H { get; set; }
 
     [JsonPropertyName("l")]
-    public decimal L { get; set; }
+    public double L { get; set; }
     
     [JsonPropertyName("c")]
-    public decimal C { get; set; }
+    public double C { get; set; }
 
     [JsonPropertyName("v")]
     public long V { get; set; }
