+++ Bot/Bot/DataSources/Csv/Models/CsvBar.cs
@@ -11,22 +11,22 @@ public record CsvBar
     public DateTime Date { get; set; }
 
     [Name("Open")]
-    public decimal Open { get; set; }
+    public double Open { get; set; }
 
     [Name("High")]
-    public decimal High { get; set; }
+    public double High { get; set; }
 
     [Name("Low")]
-    public decimal Low { get; set; }
+    public double Low { get; set; }
 
     [Name("Close")]
-    public decimal Close { get; set; }
+    public double Close { get; set; }
 
     [Name("Volume")]
     public long Volume { get; set; }
 
     [Name("Adj Close")]
-    public decimal AdjClose { get; set; }
+    public double AdjClose { get; set; }
 
     public Bar ToBotModel(string symbol) => new Bar(Date, symbol, Open, High, Low, Close, Volume, AdjClose);
 }
